app/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-celery-beat==2.5.0
django-filter==22.1
django-redis==5.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
celery==5.2.7
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.5
pydantic==1.10.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
